react-ui-debug: remove LogPanel; relocate format.test.ts (Logger lands next)
