--- packages/ui/react-ui-debug/src/components/LogPanel/index.ts
@@ -1,6 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export * from './LogPanel';
-export * from './format';
