+++ packages/ui/react-ui-debug/src/components/index.ts
@@ -2,4 +2,5 @@
 // Copyright 2026 DXOS.org
 //
 
-export * from './LogPanel';
+// Temporary: only the format helper survives until the Logger composite lands (see plan Task 9).
+export * from './Logger/format';
